Provide LLM with more return object schema context
Put some schema information into a tool. MCP Resource is likely a better
fit but not all agents currently support them (e.g. Cursor) so did it
this way instead. Thought about injecting it someplace else but GPT-5 had
an interesting suggestion, providing it as a tool allows for the schema
to be retrievable when asked as that response is explictly defined and
does not have to be derived from context if provided in an alternate
way.

I added this in repose to looking for a solution that would improve the
LLM's ability to reason about which "parent" return fields were
important for creating new VDBs for a given set of requirements.